update to_dense to utils
